Account for moving temp files across FS boundaries
os.rename only works if you're renaming across the same filesystem.
In the case where the final destination is in a different FS, you'll
get an OSError.  Now the better long term fix is to generate
the temp file on the same FS as the destination file, but for now,
we'll just use shutil.move.  This will do a byte-for-byte copy
if os.rename() fails.